# Install individual packages instead of collections